Fix broken implementation of storage key encryption
The storage key is now the hashed (and authenticated) original storage
key so that lookups succeed.  The storage key was encrypted
npn-deterministicaly in the previous implementation which made lookups
fail.

Storage keys are guaranteed not to have been tampered with (say, an
attacker replaces a stored value with another, authentic value, but that
is not related to the original storage key).